TASK08: Implemented IRQ handling task
Following files were added:
procfs_module.c - a Linux kernel module for OrangePi Zero board which switches red led light using selected interrupt.
Makefile contains target for the module's build process.
